+++ src/include/crypto/aes.h
@@ -24,11 +24,11 @@ class aes final : algorithm<aes> {
 
   ~aes();
 
-  int32_t initialize(const uint16_t mode, const uint8_t *key, const uint64_t klen, bool enable_intrinsic) noexcept;
+  int32_t initialize(const uint32_t mode, const uint8_t *key, const uint32_t klen, bool enable_intrinsic) noexcept;
 
-  int32_t encrypt(const uint8_t * const ptext, const uint64_t plen, uint8_t *ctext, const uint64_t clen) noexcept;
+  int32_t encrypt(const uint8_t * const ptext, const uint32_t plen, uint8_t *ctext, const uint32_t clen) noexcept;
 
-  int32_t decrypt(const uint8_t * const ctext, const uint64_t clen, uint8_t *ptext, const uint64_t plen) noexcept;
+  int32_t decrypt(const uint8_t * const ctext, const uint32_t clen, uint8_t *ptext, const uint32_t plen) noexcept;
 
   void clear() noexcept;
 #ifdef ENABLE_FUNCTIONS_FOR_GTEST
@@ -73,11 +73,11 @@ class aes final : algorithm<aes> {
 
   __m128i decskeys_[15];
 
-  uint32_t nr_;
+  int32_t nr_;
 
-  uint32_t nk_;
+  int32_t nk_;
 
-  uint16_t mode_;
+  uint32_t mode_;
 
   bool has_subkeys_;
