+++ src/include/crypto/algorithm.h
@@ -21,15 +21,15 @@ public:
 
   ~algorithm() {};
 
-  int32_t initialize(const uint16_t mode, const uint8_t *key, const uint64_t ksize, const bool en_intrinsic) {
+  int32_t initialize(const uint16_t mode, const uint8_t *key, const uint32_t ksize, const bool en_intrinsic) {
     return static_cast<Algorithm &>(*this).initialize(mode, key, ksize, en_intrinsic);
   }
 
-  int32_t encrypt(const uint8_t * const ptext, const uint64_t psize, uint8_t *ctext, const uint64_t csize) {
+  int32_t encrypt(const uint8_t * const ptext, const uint32_t psize, uint8_t *ctext, const uint32_t csize) {
     return static_cast<Algorithm &>(*this).encrypt(ptext, psize, ctext, csize);
   };
 
-  int32_t decrypt(const uint8_t * const ctext, const uint64_t csize, uint8_t *ptext, const uint64_t psize) {
+  int32_t decrypt(const uint8_t * const ctext, const uint32_t csize, uint8_t *ptext, const uint32_t psize) {
     return static_cast<Algorithm &>(*this).decrypt(ctext, csize, ptext, psize);
   };
