+++ src/include/crypto/camellia.h
@@ -24,11 +24,11 @@ class camellia final : public algorithm<camellia> {
 
   ~camellia() {};
 
-  int32_t initialize(const uint16_t mode, const uint8_t *key, const uint64_t klen, bool enable_intrinsic) noexcept;
+  int32_t initialize(const uint16_t mode, const uint8_t *key, const uint32_t klen, bool enable_intrinsic) noexcept;
 
-  int32_t encrypt(const uint8_t * const ptext, const uint64_t plen, uint8_t *ctext, const uint64_t clen) noexcept;
+  int32_t encrypt(const uint8_t * const ptext, const uint32_t plen, uint8_t *ctext, const uint32_t clen) noexcept;
 
-  int32_t decrypt(const uint8_t * const ctext, const uint64_t clen, uint8_t *ptext, const uint64_t plen) noexcept;
+  int32_t decrypt(const uint8_t * const ctext, const uint32_t clen, uint8_t *ptext, const uint32_t plen) noexcept;
 
   void clear() noexcept;
 #if 0
